ShopServices\ShopServices.DataAccess\Migrations\20250603065439_InitialCreate and relevant code

+++ ShopServices/ShopServices.Core/Models/EmailNotification.cs
@@ -6,7 +6,7 @@ namespace ShopServices.Core.Models
 {
     public class EmailNotification
     {
-        public uint Id { get; private set; }
+        public ulong Id { get; private set; }
         public string EmailFrom { get; private set; }
         public string EmailTo { get; private set; }
         public string Message { get; private set; }
